Apply Max's comments
